[43기 이창섭] ADD: 장바구니 존재 여부 판별 기능

+++ api/services/cartService.js
@@ -1,7 +1,19 @@
 const { cartDao } = require("../models");
+const { checkIfUserExistById } = require("../models/userDao");
+const { checkIfCartExistsById } = require("../models/cartDao");
+const { checkIfProductExistsById } = require("../models/productDao");
 const { validateQuantity } = require("../utils/validation");
 
 const createCart = async (userId, productId, color, size, quantity) => {
+  const result = await checkIfProductExistsById(productId);
+
+  if (!result) {
+    const error = new Error("NO_SUCH_PRODUCT");
+    error.statusCode = 404;
+
+    throw error;
+  }
+
   await validateQuantity(quantity);
 
   color = color.replaceAll('"', "");
@@ -15,12 +27,46 @@ const listCart = async (userId) => {
 };
 
 const updateCart = async (userId, cartId, productId, quantity) => {
+  const cartResult = await checkIfCartExistsById(cartId);
+  const productResult = await checkIfProductExistsById(productId);
+
+  if (!cartResult) {
+    const error = new Error("NO_SUCH_CART");
+    error.statusCode = 404;
+
+    throw error;
+  }
+
+  if (!productResult) {
+    const error = new Error("NO_SUCH_PRODUCT");
+    error.statusCode = 404;
+
+    throw error;
+  }
+
   await validateQuantity(quantity);
 
   return await cartDao.updateCart(userId, cartId, productId, quantity);
 };
 
 const deleteCart = async (userId, cartId, productId) => {
+  const cartResult = await checkIfCartExistsById(cartId);
+  const productResult = await checkIfProductExistsById(productId);
+
+  if (!cartResult) {
+    const error = new Error("NO_SUCH_CART");
+    error.statusCode = 404;
+
+    throw error;
+  }
+
+  if (!productResult) {
+    const error = new Error("NO_SUCH_PRODUCT");
+    error.statusCode = 404;
+
+    throw error;
+  }
+
   return await cartDao.deleteCart(userId, cartId, productId);
 };
